WIP - Move regular move creation to own module
